requirements.txt
blinker==1.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
click==8.1.7
cloudevents==1.11.0
deprecation==2.1.0
Flask==3.0.3
functions-framework==3.8.1
gunicorn==23.0.0
idna==3.9
itsdangerous==2.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mypy==1.11.2
mypy-extensions==1.0.0
packaging==24.1
python-dotenv==1.0.1
requests==2.32.3
types-requests==2.32.0.20240914
typing_extensions==4.12.2
urllib3==2.2.3
watchdog==5.0.2
Werkzeug==3.0.4
